refactor(auth): auth B.C 도메인 영역 수정 - B.C 공통 VO인 Role의 위치를 auth.domain.token.vo -> auth.domain.common.vo로 수정 - token 도메인 서비스의 dto 클래스 명 수정
