+++ src/test/java/com/simpleboard/board/auth/domain/token/service/TokenDomainServiceTest.java
@@ -5,9 +5,10 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.*;
 
-import com.simpleboard.board.auth.domain.token.dto.LoginTokenIssueCommand;
-import com.simpleboard.board.auth.domain.token.dto.RefreshTokenRotationCommand;
-import com.simpleboard.board.auth.domain.token.dto.VerifyTokenIssueCommand;
+import com.simpleboard.board.auth.domain.common.vo.Role;
+import com.simpleboard.board.auth.domain.token.dto.LoginTokenIssueParam;
+import com.simpleboard.board.auth.domain.token.dto.RefreshTokenRotationParam;
+import com.simpleboard.board.auth.domain.token.dto.VerifyTokenIssueParam;
 import com.simpleboard.board.auth.domain.token.exception.RefreshTokenEnrollBlacklistException;
 import com.simpleboard.board.auth.domain.token.exception.RefreshTokenExpiredException;
 import com.simpleboard.board.auth.domain.token.exception.RefreshTokenInvalidException;
@@ -16,7 +17,6 @@ import com.simpleboard.board.auth.domain.token.repository.TokenBlacklistReposito
 import com.simpleboard.board.auth.domain.token.util.ClockManager;
 import com.simpleboard.board.auth.domain.token.util.IdGenerator;
 import com.simpleboard.board.auth.domain.token.util.TokenProvider;
-import com.simpleboard.board.auth.domain.token.vo.Role;
 import com.simpleboard.board.auth.domain.token.vo.Token;
 import com.simpleboard.board.auth.domain.token.vo.TokenClaims;
 import com.simpleboard.board.auth.domain.token.vo.TokenPair;
@@ -87,8 +87,7 @@ class TokenDomainServiceTest {
             .build();
     when(tokenProvider.issueToken(any(TokenClaims.class))).thenReturn(dummyAccess, dummyRefresh);
 
-    LoginTokenIssueCommand cmd =
-        LoginTokenIssueCommand.builder().memberId(memberId).role(role).build();
+    LoginTokenIssueParam cmd = LoginTokenIssueParam.builder().memberId(memberId).role(role).build();
 
     // when
     TokenPair result = service.issueLoginToken(cmd);
@@ -119,7 +118,7 @@ class TokenDomainServiceTest {
             .build();
     when(tokenProvider.issueToken(any(TokenClaims.class))).thenReturn(dummy);
 
-    VerifyTokenIssueCommand cmd = new VerifyTokenIssueCommand(purpose, subject, ttl);
+    VerifyTokenIssueParam cmd = new VerifyTokenIssueParam(purpose, subject, ttl);
 
     // when
     Token result = service.issueVerifyToken(cmd);
@@ -169,8 +168,7 @@ class TokenDomainServiceTest {
             .build();
     when(tokenProvider.issueToken(any(TokenClaims.class))).thenReturn(dummyA, dummyR);
 
-    RefreshTokenRotationCommand cmd =
-        RefreshTokenRotationCommand.builder().oldRefreshRaw(oldRaw).build();
+    RefreshTokenRotationParam cmd = RefreshTokenRotationParam.builder().oldRefreshRaw(oldRaw).build();
 
     // when
     TokenPair result = service.rotateRefreshToken(cmd);
@@ -198,8 +196,7 @@ class TokenDomainServiceTest {
     when(clockManager.now()).thenReturn(now);
     when(tokenProvider.verifyAndParseToken(oldRaw)).thenReturn(oldClaims);
 
-    RefreshTokenRotationCommand cmd =
-        RefreshTokenRotationCommand.builder().oldRefreshRaw(oldRaw).build();
+    RefreshTokenRotationParam cmd = RefreshTokenRotationParam.builder().oldRefreshRaw(oldRaw).build();
 
     // when & then
     assertThatThrownBy(() -> service.rotateRefreshToken(cmd))
@@ -223,8 +220,7 @@ class TokenDomainServiceTest {
     when(clockManager.now()).thenReturn(now);
     when(tokenProvider.verifyAndParseToken(oldRaw)).thenReturn(oldClaims);
 
-    RefreshTokenRotationCommand cmd =
-        RefreshTokenRotationCommand.builder().oldRefreshRaw(oldRaw).build();
+    RefreshTokenRotationParam cmd = RefreshTokenRotationParam.builder().oldRefreshRaw(oldRaw).build();
 
     // when & then
     assertThatThrownBy(() -> service.rotateRefreshToken(cmd))
@@ -249,8 +245,7 @@ class TokenDomainServiceTest {
     when(tokenProvider.verifyAndParseToken(oldRaw)).thenReturn(oldClaims);
     when(blacklistRepository.exists("id")).thenReturn(true);
 
-    RefreshTokenRotationCommand cmd =
-        RefreshTokenRotationCommand.builder().oldRefreshRaw(oldRaw).build();
+    RefreshTokenRotationParam cmd = RefreshTokenRotationParam.builder().oldRefreshRaw(oldRaw).build();
 
     // when & then
     assertThatThrownBy(() -> service.rotateRefreshToken(cmd))
